+++ sphinx/source/changelog.rst
@@ -18,6 +18,8 @@ Changelog
 * The :ref:`isub`, :ref:`ssub`, :ref:`rsub`, :ref:`ofix`, :ref:`bfix` and
   :ref:`rfix` directives can insert an instruction after the current one
   (without first specifying a replacement for it) by using the ``+`` marker
+* :ref:`tapinfo.py` now shows pulse lengths in TZX block type 0x13 (pulse
+  sequence)
 * Added the ``CommentWidthMin`` configuration parameter for :ref:`sna2skool.py`
   (to specify the minimum width of the instruction comment field in a skool
   file)